reorganized project + implemented config plans
