+++ src/gui/app/index.html
@@ -4,7 +4,8 @@
     <meta charset="UTF-8">
     <title>Fetch</title>
     <script src="../node_modules/angular/angular.min.js"></script>
-    <script src="index.js"></script>
+    <script src="fetchController.js"></script>
+    <script src="trayFactory.js"></script>
     <link rel="stylesheet" href="styles.css">
   </head>
   <body ng-app="fetchApp" ng-controller="fetchController as vm">
