+++ src/gui/app/trayFactory.js
@@ -1,0 +1,27 @@
+(function(){
+	const { remote } = require('electron');
+	const { Tray } = remote;
+	
+	let app = angular.module('fetchApp');
+	
+	app.factory('trayFactory', trayFactoryWrapper);
+	
+	function trayFactoryWrapper() {
+		
+		function trayFactory() {
+			let tray = new Tray('app/res/fetch.png');
+			tray.setTitle('Fetch');
+			tray.setToolTip('Fetch');
+
+			this.tray = tray;
+		}
+
+		angular.extend(trayFactory.prototype, {
+			buildMenu: function() {
+				console.log('building menu');
+			}
+		})
+	
+		return trayFactory;
+	}
+})();
