+++ src/gui/config.js
@@ -1,31 +1,46 @@
 const fs = require('fs');
 const path = require('path');
 
-module.exports = getConfig();
+module.exports = {
+	get: get,
+	put: put
+}
 
-function getConfig() {
-	let configPath;
+function getConfigPath() {
+	let dir;
+
 	if (process.env.HOME) {
 		// ubuntu machine
-		configPath = path.join(process.env.HOME, '.fetch', 'config.json');
+		dir = path.join(process.env.HOME, '.fetch');
 	} else if (process.env.LOCALAPPDATA) {
 		// windows machine
-		configPath = path.join(process.env.LOCALAPPDATA, 'Fetch', 'config.json');
+		dir = path.join(process.env.LOCALAPPDATA, 'Fetch');
 	}
+	return { dir: dir, file: 'config.json' };
+}
 
-	// load default config
-	let config = JSON.parse(fs.readFileSync('default_config.json'));
-	
-
-	// return saved config if exists
-	if (fs.existsSync(configPath)) {
-		let savedConfig = JSON.parse(fs.readFileSync(configPath));
-	
-		// replace missing config with defaults
-		for (let key in savedConfig) {
-			config[key] = savedConfig[key];
-		}
+function get() {
+	let configPath = getConfigPath();
+	let configFile = path.join(configPath.dir, configPath.file);
+	if (!fs.existsSync(configFile)) {
+		save(JSON.parse(fs.readFileSync('misc/default_config.json')));	
 	}
+	return JSON.parse(fs.readFileSync(configFile));
+}
 
-	return config;
+function put(kv) {
+	let config = get();
+	for (let key in kv) {
+		config[key] = kv[key];
+	}
+	save(config);
+}
+
+function save(config) {
+	let configPath = getConfigPath();
+	if (!fs.existsSync(configPath.dir)) {
+		fs.mkdirSync(configPath.dir, { recursive: true });
+	}
+	let configFile = path.join(configPath.dir, configPath.file);
+	fs.writeFileSync(configFile, JSON.stringify(config));
 }
