+++ src/gui/misc/default_config.json
